backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
python-dateutil>=2.8.2
requests>=2.31.0
gunicorn>=21.0.0
whitenoise>=6.0.0
